Finish 1.0.0 voting with Karl, so we can start rolling the beta tarball.
Push r8708 into 1.0.1; it didn't have any votes, which means we can't
put it in the beta tarball, which is supposed to be "identical" to the
real tarball.  Both of us don't feel ready to +1 it, because we're not
sure whether the 5.6.0 perl requirement is controversial or not, and
we haven't followed the mailing list thread carefully.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/1.0.x@848854 13f79535-47bb-0310-9956-ffa450edef68