Locking branch:  obscure BDB changes-table bugfix, done with cmpilato.
Normal clients don't stimulate this bug, but the new lock-checking
behavior in svn_fs_commit_txn() stimulated it in changes-test #5.  The
locking branch should now pass regression tests again.

In particular, changes-test #5 was deleting a child of a dir, then
deleting the parent dir, all in one transaction.  When
svn_fs_commit_txn() started looping over over the list of txn changes
(to enforce any locks), it was getting back a mutable-id for the
deleted parent dir... a mutable-id of a node that was now gone!  

* subversion/libsvn_fs_base/tree.c 
  (txn_body_delete):  don't write a mutable node-id into the changes list.
                      Instead, write the immutable predecessor id.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/locking@853030 13f79535-47bb-0310-9956-ffa450edef68