First crack at permitting files to be replaced.  This doesn't handle properties yet which makes it incomplete.
The main problem is that when we add a node with history we set the base copies
(properties and text) as they are from the copy.  This allows diff to show the
changes from the copy and allows us only to transmit the delta of any changes
from the copy (if any at commit time).  

This runs into problems when adding a node with history which is replacing an
existing versioned node that has been deleted.  We need the text base in order
to revert the delete/replace from the original file that is our BASE.  But
we want the text base of the copy for the reasons given above.

The solution is to copy the base files to another location before replacing
them with the new base files.  If the change is committed we remove this
extra file.  If the change is reverted we use it instead of the normal
base files to revert it.

* subversion/libsvn_wc/copy.c
  (copy_file_administratively): Detect when we're replacing a versioned file
    and copy the text base to another file.

* subversion/libsvn_wc/wc.h
  (SVN_WC__REVERT_EXT): New constant for the file name extension for the
    replaced files base.
  (SVN_WC__VERSION): Bump because this changes our on disk wc format.

* subversion/libsvn_wc/adm_ops.c
  (svn_wc_process_committed): Clean up the replaced files base file.
  (svn_wc_add): Copies have no checksum, neither do replaced files.
  (revert_admin_things): Remove copy flags when reverting a replacement.
    When actually doing the revert prefer the replaced base over the copied
    base.

* subversion/libsvn_wc/adm_files.c
* subversion/lbisvn_wc/adm_files.h
  (svn_wc__text_revert_path): New function to get the path to the replaced
    base.
  
* subversion/libsvn_wc/update_editor.c
  (install_file): Look for replaced files and treat them as changed even if
    there is no change from their text base.  Also don't write the checksum
    on replaced files.



git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/wc-replacements@853386 13f79535-47bb-0310-9956-ffa450edef68