On the fs-atomic-renames branch:
Introduce an alias record in the copies table, with the sole purpose of
allowing us to refer back to a previous entry's data without introducing
the possibility of introducing primary key conflicts by reusing ids.

Once again, this is not the final word on this stuff, but it does seem
to get the tests passing, so it seems like a reasonable place to start.

* subversion/libsvn_fs_base/tree.c
  (make_path_mutable): Update comment, handle fixed inheritance by
   inserting an alias record.

* subversion/libsvn_fs_base/bdb/copies-table.c
  (svn_fs_bdb__get_copy): Dereference alias records to get the copy
   they point to.

* subversion/libsvn_fs_base/fs.h
  (copy_kind_t): Add a copy type for aliases.

* subversion/libsvn_fs_base/util/fs_skels.c
  (is_valid_copy_skel,
   svn_fs_base__parse_copy_skel,
   svn_fs_base__unparse_copy_skel): Handle aliases.

* subversion/tests/libsvn_fs/fs-test.c
  (move_plus_copy_test): Tweak the expected results to match reality.
  (test_funcs): Mark move_plus_copy_test as PASS.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/fs-atomic-renames@859070 13f79535-47bb-0310-9956-ffa450edef68