Revert r21067, that is, go back to the depth names introduced in r21035.
After working with both the original WebDAV names and the new LDAP
names for a while, I found it difficult to think with the LDAP names,
and kept wanting to use the WebDAV ones.  IMHO this is not because of
familiarity, but rather because the LDAP names are subtly misleading,
whereas the WebDAV names don't get in the way of their own meanings.
See also Max's mail, which I now agree with:

   http://subversion.tigris.org/servlets/ReadMsg?list=dev&msgNo=118707
   From: Max Bowsher <maxb1@ukf.net>
   To: "C. Michael Pilato" <cmpilato@collab.net>
   Cc: Peter Samuelson <peter@p12n.org>,
       Karl Fogel <kfogel@red-bean.com>, 
       dev@subversion.tigris.org
   Subject: Objection to change in svn_depth_t component naming (r21067)
            (was: Re: [PATCH] Infrastructure changes for incomplete
            directory support.)
   Date: Mon, 14 Aug 2006 11:24:27 +0100
   In-Reply-To: <44DB3D2F.7000907@collab.net>
   Message-ID: <44E04F5B.8000205@ukf.net>

And the thread following from it:

http://subversion.tigris.org/servlets/BrowseList?list=dev&by=thread&from=495571


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/incomplete-directories@861241 13f79535-47bb-0310-9956-ffa450edef68