Abort the test script if an individual test fails.
The tests in this file build upon one another -- a checkout, followed by
an add, followed by a commit, and so on.  If the checkout fails there's no
point continuing the rest of the test; it just leads to spurious errors
when the add and commit both fail.  Use the Fatal module to cause the
testing functions, and open() and close(), to die() instead of returning
false.

* subversion/bindings/swig/perl/native/3client.t: Use Fatal to have open,
  close, ok, is, and isa_ok die() instead of returning false.  Remove the
  "... or die '...'" stanza from places where it's used.



git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/perl-bindings-improvements@863316 13f79535-47bb-0310-9956-ffa450edef68