On the merge-tracking branch: Fix svn_rangelist_merge from r23600.
First off r23600 was seriously broken.  I wanted to merge the rangelists in
place with no third array, but with the need to keep the resulting rangelist
sorted and avoid duplicate ranges it became quite convoluted.  So reverted
to an implementation not that different from r23592 while still keeping the
simpler API. 

* subversion/libsvn_subr/mergeinfo.c
  (svn_rangelist_merge): Use a third array internally.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/merge-tracking@863811 13f79535-47bb-0310-9956-ffa450edef68