Make depth parameters optional in the svn:// protocol in a cleaner way, and document the depth parameters in the first place (as should have done in r23098 or earlier).
Suggested by: malcolm

Note: This is similar to r23760, but the reason lundblad didn't do
these as well in r23760 is that these depth parameters will probably
go away when we implement lundblad's suggestion about how to set
default depth in reporters:

   http://subversion.tigris.org/servlets/ReadMsg?list=dev&msgNo=124363
   From: plundblad@google.com (Peter Lundblad)
   To: dev@subversion.tigris.org
   Subject: [sparse-directories] Default depth in reporter?
   Date: Wed, 7 Mar 2007 11:44:47 +0100
   Message-ID: <17902.38815.95707.550653@amazonas.zrh.corp.google.com>

But, one step at a time.  These are easy corrections, and while I'm
now pretty sure about the "set-path and link-path are all we need"
approach, I'm not 100% positive yet.  In the meantime, bad protocol
is the devil's playground, so let's make it nicer.

* subversion/svnserve/serve.c
  (update, switch_cmd, status, diff): Make depth an optional raw word,
    not an optional tuple with an optional word inside.

* subversion/libsvn_ra_svn/client.c
  (ra_svn_update, ra_svn_switch, ra_svn_status, ra_svn_diff): Adjust
    accordingly.

* subversion/libsvn_ra_svn/protocol: 
  (3.1.1. Main Command Set): Document depth parameter


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/sparse-directories@864004 13f79535-47bb-0310-9956-ffa450edef68