[This revision finishes the work started in r27846; this log message covers the changes in both revisions.]
On the sqlite-node-origins branch:

Abstract sqlite functionality that is not specific to mergeinfo out of
mergeinfo-sqlite-index.c and into a new sqlite-util.c.

Create an empty node_origins table (but don't use it yet); this
involved writing code to upgrade a db containing mergeinfo tables but
no node_origin table to have both.  While this is not strictly
necessary for compatibility (because 1.5 is not released yet), this
will both make lives easier for developers now *and* give us an
infrastructure for upgrading the database in future versions.

* subversion/libsvn_fs_util/sqlite-util.c
  New.  Move a bunch of functions here (renaming them to
  svn_fs__sqlite_*)...

* subversion/libsvn_fs_util/mergeinfo-sqlite-index.c
  ... from here.

* subversion/libsvn_fs_util/mergeinfo-sqlite-index.h
  ... and declare them here.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/sqlite-node-origins@867921 13f79535-47bb-0310-9956-ffa450edef68