Reverse-merge and add -0 to r31357.
Serf tests are failing on OSX with this revision applied.
All tests, including bindings, pass with this revision backed
out.

I am not against having this revision, I just do not want to
block the release for a non-crucial fix that is causing the tests
to fail (for at least me).

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/1.5.x@871503 13f79535-47bb-0310-9956-ffa450edef68