On the file externals branch.
File externals now place into the working copy a file.  Lots of work
to do, but the file now appears just as a switched file.  Removing the
external from the svn:externals now removes it from the working copy
if there are no modifications to it, otherwise it leaves it with mods.

* subversion/libsvn_client/externals.c
  (switch_file_external):
    New function that is very similar to switch_dir_external(), except it
    creates an empty file and adds it into a working copy.  If the
    file already exists, just switch to the proper URL.
  (handle_external_item_change):
    For each of the major code paths, checking out/exporting, updating
    and removing, handle a file external.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/file-externals@872766 13f79535-47bb-0310-9956-ffa450edef68