* STATUS: Vote for the issue #3275 fixes.
While I don't have a proxy server to test this with, and I'm not super-
familiar with the mod_dav_svn code, I think I've given this due diligence.
The logs and diff looked good, and the comments in the issue tracker seemed
to indicate that this is the correct fix.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/1.5.x@874802 13f79535-47bb-0310-9956-ffa450edef68