On the 1.5.x-r30756 branch:
Backport of r30756, r32907, and r32913 to 1.5.x.
Allow 'svn merge' to merge diffs which add 'svn:eol-style' properties
to files that used to have inconsistent newlines.

As part of conflict resolution, also add some bits from r32405:
* subversion/tests/cmdline/merge_tests.py
  (local_path, svn_commit, svn_merge, svn_propset): New helper
   functions added in r32405, needed by the test added in r32907.
There is no mergeinfo record of this change, because r32405 is
too large to be merged entirely.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/1.5.x-r30756@874857 13f79535-47bb-0310-9956-ffa450edef68