On the 1.5.x-reintegrate-improvements branch, some mergeinfo-fu:
In a WC for the 1.5.x-reintegrate-improvements branch I performed these 
four commands:

>svn merge
 http://svn.collab.net/repos/svn/branches/1.5.x/subversion/include/private
 subversion\include\private  -r34511:34796 --ignore-ancestry --depth empty
--- Merging r34512 through r34796 into 'subversion\include\private':
 G   subversion\include\private

>svn merge
 http://svn.collab.net/repos/svn/branches/1.5.x/subversion/include/private
 subversion\include\private --record-only -r34511:34796

>svn merge
 http://svn.collab.net/repos/svn/branches/1.5.x/subversion/libsvn_subr
 subversion\libsvn_subr -r34511:34796 --ignore-ancestry --depth empty
--- Merging r34512 through r34796 into 'subversion\libsvn_subr':
 G   subversion\libsvn_subr

>svn merge
 http://svn.collab.net/repos/svn/branches/1.5.x/subversion/libsvn_subr
 subversion\libsvn_subr -r34511:34796 --record-only

Why?  So we can use --reintegrate when merging
 1.5.x-reintegrate-improvements back to trunk.  In r34970 I --record-only
merged r34766-34768 from 1.5.x to 1.5.x-reintegrate-improvements.  But
r34768 in 1.5.x added explicit mergeinfo to subversion\libsvn_subr and
subversion\include\private, and since I only did a --record-only merge of
that revision those two subtrees never got any explicit mergeinfo until
the *next* time I did a "real" merge from of r34769-34770 from 1.5.x to
1.5.x-reintegrate-improvements in r34792 -- In r34792 the two subtrees 
finally get explict mergeinfo but this mergeinfo only represented the diff
between 1.5.x@34768 and 1.5.x@34769:

>svn diff --depth empty -r34768:34770
 http://svn.collab.net/repos/svn/branches/1.5.x/subversion/include/private 

Property changes on: .
___________________________________________________________________
Modified: svn:mergeinfo
   Merged /trunk/subversion/include/private:r30257,33024,33026


>svn pg svn:mergeinfo -v
 http://svn.collab.net/repos/svn/branches/1.5.x-reintegrate-improvements/ \
 subversion/include/private@34792
Properties on '
http://svn.collab.net/repos/svn/branches/1.5.x-reintegrate-improvements/
subversion/include/private':
  svn:mergeinfo
    /trunk/subversion/include/private:30257,33024,33026

>svn diff --depth empty -r34768:34770
 http://svn.collab.net/repos/svn/branches/1.5.x/subversion/libsvn_subr 

Property changes on: .
___________________________________________________________________
Modified: svn:mergeinfo
   Merged /trunk/subversion/libsvn_subr:r30257,33024,33026
   
>svn pg svn:mergeinfo -v
 http://svn.collab.net/repos/svn/branches/1.5.x-reintegrate-improvements \
 /subversion/libsvn_subr@34792
Properties on '
http://svn.collab.net/repos/svn/branches/1.5.x-reintegrate-improvements/
subversion/libsvn_subr':
  svn:mergeinfo
    /trunk/subversion/libsvn_subr:30257,33024,33026

This is a problem for reintegrate because it appears that 
subversion\libsvn_subr and subversion\include\private are missing 
revisions r34512-34770:

>svn merge --reintegrate
http://svn.collab.net/repos/svn/branches/1.5.x-reintegrate-improvements .
svn: Reintegrate can only be used if revisions 34512 through 34801 were
previously merged from http://svn.collab.net/repos/svn/branches/1.5.x to
the reintegrate source, but this is
 not the case:
  branches/1.5.x-reintegrate-improvements/subversion/include/private
    Missing ranges: /branches/1.5.x/subversion/include/private:34512-34770
  branches/1.5.x-reintegrate-improvements/subversion/libsvn_subr
    Missing ranges: /branches/1.5.x/subversion/libsvn_subr:34512-34770

Still here?  Good for you!  Anyway, the --ignore-ancestry merges at depth 
empty add the "missing" mergeinfo that the --record-only merge would have 
added.  This is not really necessary to get reintegrate to work, but it 
will prevent the reintegrate merge from removing a whole slew of mergeinfo 
on the two subtrees in question.  The --record-only merges are what allow 
the reintegrate to actually work, by adding the missing blocks of revisions.

For those of you with no lives or who really pay attention to merge 
tracking bugs (are those different?) you'll notice that this whole problem 
is a lot like the one solved by r32975 'Fix bug which occurs when a merge 
adds explicit mergeinfo to a subtree that had no explicit mergeinfo prior 
to the merge.', but with --record-only merges.  I say "problem" because I 
don't see this particular situation as a bug, more as a warning to be 
careful what you do with --record-only.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/1.5.x-reintegrate-improvements@874876 13f79535-47bb-0310-9956-ffa450edef68