Reduce the average number of iterations required by svn_io_open_uniquely_named() and reduce the runtime of each iteration.
Often, the initial temp. path suggested to the function is very generic one. If as a 
result of a crash, other failure or simple parallel execution, many temporary files
share the same prefix causing the algorithm to take many iterations to find an unused
(prefix, iteration, suffix) combination. Just throw an additional random element in
there and we are usually good after the second iteration.

* subversion/libsvn_subr/io.c
  (svn_io_open_uniquely_named): add a random element to the path once the first
  test failed; don't call UTF8 conversion  unnecessarily

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/performance@985472 13f79535-47bb-0310-9956-ffa450edef68