Fix another packing issue. Due to concurrent access alone it may happen that between determining the rev file name and actually opening it, that very revision may get packed (actually, get deleted after packing).
The file handle cache adds another issue: an open file handle 
may not be suitable due to different open flags and the file 
must be re-opened. While the existing handle would have
kept the file content alive, opening a new handle to the same
file will fail if the file was deleted before. 

Since there can be only one transition from non-packed to
packed, we need to retry only once if we could not find the
revision file in question.

* subversion/libsvn_fs_fs/fs_fs.c
  (open_pack_or_rev_file): retry once because the file might have
   gotten packed.

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/performance@1029381 13f79535-47bb-0310-9956-ffa450edef68