On the 1.7.x-issue3991 branch, merge the issue #3991 fix from trunk, resolving conflicts.
This fix depends on a piece of the issue #3814 fix. This bit has been
merged in via a subtree merge to libsvn_client/patch.c. As a result
this branch fixes issue #3991 independently of the 1.7.x-issue-3814
branch but might conflict with the issue #3814 branch depending on
which branch gets reintegrated into 1.7.x first.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/1.7.x-issue3991@1206730 13f79535-47bb-0310-9956-ffa450edef68