On the ev2-export branch: Use an Ev2-style driver to handle repos->repos copies.
Part of this commit is rather bogus, namely the bit where we still do a
delete+add.  This *should* be a move, and will be adjusted to such in a
future commit, but the fact that it doesn't cause any tests to fail as
currently implemented is somewhat strange (the Ev2 internal checks should
catch this, methinks).

* subversion/libsvn_client/copy.c
  (path_driver_cb_baton, path_driver_cb_func): Remove.
  (fetch_remote_kind_func, fetch_remote_props_func, drive_single_path,
   drive_editor): New.
  (repos_to_repos_copy): Delegate the driving of editors to the helper
    function.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/ev2-export@1310005 13f79535-47bb-0310-9956-ffa450edef68